Moved Botvac functions to a library
